requirements.txt
torch==1.13.1+cu117
numpy==1.21.5
scipy==1.6.0
blobfile==2.0.1
Pillow==9.0.1
piq==0.7.0
six==1.15.0
scikit-learn==1.0.2
h5py==3.8.0
tensorflow==2.11.0
torchvision==0.14.1+cu117
opencv-python==4.5.5.62
matplotlib==3.3.2